(41) Define SCOPE_PID env variable for general use.
