+++ src/dbg.c
@@ -262,14 +262,3 @@ scopeLog(const char *msg, int fd, cfg_log_level_t level)
     return;
 }
 
-int
-checkEnv(char *env, char *val)
-{
-    char *estr;
-    if (((estr = getenv(env)) != NULL) &&
-       (strncmp(estr, val, strlen(estr)) == 0)) {
-        return TRUE;
-    }
-    return FALSE;
-}
-
