+++ src/javaagent.c
@@ -1,4 +1,5 @@
 #include "dbg.h"
+#include "fn.h"
 #include "os.h"
 #include "state.h"
 #include <jni.h>
@@ -623,7 +624,7 @@ initJavaAgent() {
         buf = malloc(bufsize);
         snprintf(buf, bufsize, "%s%s%s", env != NULL ? env : "", env != NULL ? " " : "", opt);
 
-        int result = setenv("JAVA_TOOL_OPTIONS", buf, 1);
+        int result = g_fn.setenv("JAVA_TOOL_OPTIONS", buf, 1);
         if (result) {
             scopeLog("ERROR: Could not set JAVA_TOOL_OPTIONS failed\n", -1, CFG_LOG_ERROR);
         }
